Tests
test


test


test

+++ Poseiden-skeleton/src/test/java/com/nnk/springboot/services/RuleNameServiceTest.java
@@ -1,10 +1,6 @@
 package com.nnk.springboot.services;
 
-import com.nnk.springboot.domain.BidList;
-import com.nnk.springboot.domain.CurvePoint;
 import com.nnk.springboot.domain.RuleName;
-import com.nnk.springboot.domain.Trade;
-import org.junit.Rule;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -26,8 +22,10 @@ public class RuleNameServiceTest {
     @BeforeEach
     public void setUp(){
         List<RuleName> ruleNames = ruleNameService.findAll();
-        for (RuleName ruleName: ruleNames) {
-            ruleNameService.delete(ruleName.getId());
+        if (ruleNames.size() !=0){
+            for (RuleName ruleName: ruleNames) {
+                ruleNameService.delete(ruleName.getId());
+            }
         }
     }
